Fix PIOv4 driver not to access out of bounds memory.
The driver used periphal IDs as offset of a base address to access to
the desired PIO group. Since these ids doesn't match the memory mapping,
random access and writes were applied.

The "id" field of _pin struct is now associated to a PIO group
number for the v4 PIO API.