peripherals/mcan: resolve potential RAM caching issues
Allocate the Message RAM from non-cached memory.

The Buffer Elements in the Message RAM are contiguous and not aligned on
cache lines. If caching was turned on, it would be necessary to
explicitly clean the data cache lines matching transmit buffers, and
then invalidate the data cache lines matching receive buffers. Since
these sections overlap in the cache, strong constraints would apply as
soon as several transmit and receive buffers were used concurrently.

Signed-off-by: Aubin Constans <aubin.constans@atmel.com>