peripherals/sdmmc: create driver
The SDMMC driver provides the interface to configure and use the SDMMC
peripheral.
The SDMMC peripheral has been designed to comply with the the SD Host
Controller Specification (version 3.00 on Atmel SAMA5D2x MPUs).
In the terms of the Specification, the SDMMC peripheral is referred to
as the Host Controller, whereas the SDMMC driver is the Host Driver.

The SDMMC driver is meant to be accessed via the SD/MMC library
(libsdmmc), which implements the most part of both MMC and SD physical
layers, and abstract differences between these devices.

A bit of history:
- create HAL for libsdmmc
- resolve data alignment fault when memcpying to double word
- add support for SDMMC_IOCTL_SET_CLOCK
- add support for SDMMC_IOCTL_GET/SET_HSMODE
- add support for SDMMC_IOCTL_SET_BUSMODE
- skip command index errors
- add support for SDMMC_IOCTL_POWER and SDMMC_IOCTL_RESET
- fix endless loop when resetting the e.MMC
- add support for SDMMC_IOCTL_CANCEL_CMD
- fix verification of R2, R3 and R4 command responses
- reorder response bytes for libsdmmc compatibility
- configure data timeout
- disallow command index errors again
- also support TRFC without prior CMDC interrupt
- fix frequency of device clock
- upon data timeout return SDMMC_ERR_IO
- add auto SET_BLOCK_COUNT command
- reply to SDMMC_IOCTL_GET_XFERCOMPL positively
- add auto STOP_TRANSMISSION command
- add SDMMC_IOCTL_GET_DEVICE
- handle potential errors in auto STOP_TRANSMISSION response
- add support for Advanced DMA 2
- add SDMMC_IOCTL_GET_BUSMODE
- fix detection of card presence
- address premature Transfer Complete status

Signed-off-by: Aubin Constans <aubin.constans@atmel.com>