examples/sdmmc_sdcard: create example
Example of using the SDMMC driver on the SAMA5D2 Xplained Ultra evaluation
board.

The example provides access to the e.MMC (SDMMC0) and to the SD or MMC card
inserted in the on-board connector (SDMMC1).
Please note that due to software limitations, it is not possible currently
to access both slots at the same time.

A bit of history:
- start using libsdmmc
- fix total size of memory device
- display extended CSD in case of MMC device
- customize FatFs configuration
- build FatFs library
- add RAW read/write test, and FAT mount & lsdir test
- fix EWARM build
- let FatFs call SD_Init
- enable DMA
- add runtime option to use DMA
- align FatFs buffers on cache lines
- align libsdmmc data buffers on cache lines
- fix allocation of large buffers

Signed-off-by: Aubin Constans <aubin.constans@atmel.com>