libsdmmc: upgrade multiple-block data transfer
Devices not supporting the SET_BLOCK_COUNT command require the host to
send the STOP_TRANSMISSION command instead, when the last data block has
just been transferred.
As STOP_TRANSMISSION is timing dependent, if it's triggered by software
then the time window is likely to be missed.
Peripherals compatible with the SD Host Controller Specification - such
as SDMMC, but not HSMCI - address this constraint in hardware with the
Auto CMD12 capability.

Do not issue STOP_TRANSMISSION before the transfer but after. Let the
driver do it automatically, if supported.

If the device or the peripheral detects an error during the transaction,
it becomes the library's duty to issue the STOP_TRANSMISSION command. If
the device does not respond to STOP_TRANSMISSION, the host shall attempt
a SEND_STATUS command to recover.
Mind that upon STOP_TRANSMISSION, the device is likely to temporarily
switch to the Programming State.

Remove the bState flag in struct _SdCard. Not used anymore, as the data
transfer is now stopped upon return of SD_Read or SD_Write.

Change the meaning of the bStatus field in struct _SdCard: now used to
hold the code of an error that occurred in this session and could not be
recovered.

Signed-off-by: Aubin Constans <aubin.constans@atmel.com>