target/sama5d{2,4}: prevent IAR from initializing .region_{sram,ddr,ddr_nocache} sections
Add these 3 sections to the 'do not initialize' ILINK directive.
Motivations:
1. When running the program from the internal SRAM, the low-level, early
initialization script does not configure the DDR-SDRAM. Hence early
access to the DDR-SDRAM to initialize sections of the memory array
simply can not happen.
2. Align IAR ILINK with GNU LD. Our GNU linker scripts flag these
sections with the NOLOAD directive, which at runtime implies the program
loader will not load these sections into memory.

Warning: ICC still considers the sections below as zero-initialized, by
default. May included variables be assigned a null initializer, no
warning is raised.
The variables may be assigned the '__no_init' ICC-specific keyword to
fix this.
Same warning with GCC, as it will not warn the programmer if a variable
belonging to an unloaded section is assigned an initializer.

Walked through the tree and verified that no variable placed in
.region_{sram,ddr,ddr_nocache} uses any initializer.

A symptom that was observed:
Running from internal SRAM and allocating more than 4 MiB from
DDR3-SDRAM, the UART transmitter was not properly initialized.
The variables placed in DDR3-SDRAM were listed in the INIT TABLE section
of the map file.

Signed-off-by: Aubin Constans <aubin.constans@atmel.com>