peripherals/sdmmc: support swapped Command Complete and Transfer Complete
There are cases, notably CMD7 with address and R1b, where the Transfer
Complete interrupt precedes Command Complete. In such cases, the
command/response is still in progress, we shall wait for Command Complete.

The Command Complete event was received upon processing of the subsequent
command, and of course was invalid in this context.

Signee-off-by: Aubin Constans <aubin.constans@atmel.com>