libsdmmc: remove global variable
Remove the 'status1' global variable. As the device status is read by the
CPU from the response register (which is referred to as direct data read),
the data memory cache can be enabled for this variable as well.

Also, without any global variable now, several instances of the library
can be used simultaneously.

Signed-off-by: Aubin Constans <aubin.constans@atmel.com>