peripherals/hsmcid: fix SDIO byte and block I/O
Fix a violation of the libsdmmc API. When transferring a data chunk
(vs transferring data blocks), sSdmmcCommand::wNbBlocks shall be set to 1,
and sSdmmcCommand::wBlockSize shall be set to the size of the chunk,
in bytes.

The special case of SDIO byte transfer on the HSMCI peripheral shall
be abstracted by the driver.

Also, address an apparent confusion between SDIO byte I/O and the role
of the FBYTE flag in the HSMCI Mode Register.

Signed-off-by: Aubin Constans <aubin.constans@atmel.com>