Makefiles: Graceful raise an error when TARGET or VARIANT is not supported
In an example Makefile, it is possible to define the supported varaints
or the supported target. The makefile varaibles AVAILABLE_VARIANTS and
AVAILABLE_BOARDS are exported to user now.

When declaring one of those 2 variables in an example Makefile, you are
supposed to list the supported targets or the supported variants for
this example.

Signed-off-by: Laurent Fazio <laurent.fazio@atmel.com>