peripherals/pmc: fix pmc_configure_gck & friends
The pmc_configure_gck function was not changing PMC_PCR:GCKCSS to the
requested clock source, but to a potentially invalid value.
The issue was observed typically when using CLASSD, ISC, and SDMMC
peripherals, as their default clock source is not the slow clock.

Refer to RFOSAMAV-447 and RFOSAMAV-495 tracking issues.

Signed-off-by: Aubin Constans <aubin.constans@atmel.com>