drivers/usb: Fix a bug in usbd_hal_unhalt API.
It is from a case for SAMA5D3, the number is 0026549.
SAMA5D2/Canopus has the same problem.
Description:
When printing a doc on Mac for the first time, no data lost.
Then print the second doc, the first 71 bytes will be lost on the
terminal but exist on the USB line.
When unhalt the endpoint, data toggle does not reinitialized which result
in data lose.

Solution:
The following words are from usb spec.
For endpoints using data toggle, regardless of whether an endpoint has
the Halt feature set, a ClearFeature(ENDPOINT_HALT) request always results
in the data toggle being reinitialized to DATA0.

Signed-off-by: tingting.lv <tingting.lv@SHALT0343.corp.atmel.com>