peripherals/hsmcid: support command response with busy signal
The busy signal that may follow the R1b command response was ignored.

For instance, when testing selected MMCplus cards, further to the SWITCH
command (switching to High Speed timing mode, or to 8-bit data bus width),
the subsequent SEND_STATUS command was finding that the card was still
busy in the Programming State.

Signed-off-by: Aubin Constans <aubin.constans@atmel.com>