drivers/peripherals/dma: Add generic DMA API drivers.
Implement a generic DMA API in order to support the DMAC and XDMAC
peripherals present on the multiple architecture we are supporting.

The API does not allow to use all the advanced feature fo the XDMAC, but
it is a practical tool to implement our example. The XDMA example still
exists to demonstrate the advanced features.

Signed-off-by: LI BIN <bin.li@atmel.com>