drivers/peripherals/twi: Implement a scatter/gather mechanism
In order to implement a scatter/gather mechanism to send data originating
from different buffers, it is mandatory to control START/STOP conditions
precisely. To reach that goal, the use of IADDR will be deprecated.

This modification is required to prepare the introduction of TWI bus
driver.

Currently, only AT24 driver has been ported to the new API.

Signed-off-by: Laurent Fazio <laurent.fazio@atmel.com>