samba_applets/common: add workaround for stack pointer issue on older devices
On SAM9, the romcode does not handle its stack pointer correctly. After each
call to the applet, the stack pointer is not reset correctly, leading to a
stack overflow after several applet calls.

The workaround is to save the rom-code stack pointer on the first applet call
and to restore this stack pointer value after each subsequent applet call.

Signed-off-by: Loic Lefort <loic.lefort@atmel.com>