drivers/peripherals/dma: Add DMA linked list pool management
This commit adds a DMA linked list pool management in order to allocate
a linked list elemennt from the pool as we do not allow malloc. When used,
the linked list element return to the pool of LL elements.

This pool simplifies the dynamic insertion of linked list element to be
transferred by the DMA.

Signed-off-by: LI BIN <bin.li@atmel.com>