target/common: add new chip_pins_flexcom.h shared header
Flexcom definitions are huge and the same for all devices having Flexcom, move
them to a separate, shared, header file.

Additionally, since definitions in this header file are all very similar, a
script was created to automatically generate the header file.  This tool is
only needed when the header needs to be re-generated (for example to fix the
template or to increase in number of Flexcom devices or IOSet supported.)

Signed-off-by: Loic Lefort <loic.lefort@atmel.com>