utils/timer: fix short delay in usleep()
From @dandersonee on github.com:
I have seen a few cases where code gets stuck in usleep during driver
initialization. Usually when running under debugger or with caches
disabled. This happens with small values passed to usleep.
The debugger confirmed that the timer counter is still counting but
usleep is waiting for TC_SR_CPCS but the counter has already passed the
compare value.
What seems to be happening is that by the time usleep writes the compare
value to rc, the timer counter has already advanced past the compare value
and usleep sits waiting for TC_SR_CPCS that is not coming for a very long
time.

Signed-off-by: Laurent Fazio <laurent.fazio@atmel.com>