drivers/peripherals/pmc: fix pmc_set_main_oscillator_freq()
The pmc_set_main_oscillator_freq() function changes the source of the MAIN
clock: this has a side effect on the PLLA since the MAIN clock is the
source of the PLLA. The PLLA is no longer working correctly.

If the PLLA is enabled, this patch disables the PLLA before changing the
MAIN clock then enables again the PLLA once the final configuration of the
MAIN clock has been set.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>