target/samv71: lower the PLLA multiplier
This will lower the PLLA Clock (PLLACK) to (24+1)_12=300MHz
instead of (49+1)_12=600MHz. Also change the PMC_MCKR prescaler
to 1 instead of 2 which will result in the same HCLK and MCK
as before this change.

This will save about 6mA during idle and also bring the
PLLACK within specification (max is 500MHz according to
Table 58-26, page 2055 in datasheet DS60001527B).

Signed-off-by: Johannes Carlsson <johannes.carlsson@acconeer.com>