Revert "drivers/sdmmc: fix the conditions checked for MMC's HS200 mode support"
This reverts commit 4ef9b12dcac702cda9b778c41654e3d171106c13.

When the customer board is designed for the HS200 bus speed mode, the
board-specific function board_cfg_sdmmc() shall be modified so it sets
the correct SDMMC capabilities.
The logic of the generic driver then automatically recognizes if the
capabilities comply with HS200 requirements.

We use the Voltage Support capabilities to indicate the supported
signaling level(s) (VCCQ).
In the case of e.MMC devices (vs SD Cards), changing VCCQ on the fly is
illegal. The HS200 bus speed mode requires that the board supplies the
e.MMC with 1.8V VCCQ right from e.MMC power-on time and onwards.
Accordingly, the Voltage Support capabilities shall indicate that the
signaling level (VCCQ) is 1.8V and not 3.0/3.3V.