examples/sdmmc_sdcard: add support for sam9x60-ek board
Configure the SD Card slot, which is driven by the SDMMC0 Controller.

Change the divider of the Generated Clock used with SDMMCx.
If the PLLA is either 600MHz or 500MHz, the Generated Clock divider should
be at least 6 so MULTCLK is lower than 100MHz, hence BASECLK is lower than
50MHz. This way, both the BASECLKF and TEOCLKF bitfields don't overflow in
the SDMMC_CA0R Capability Register.
Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>

Signed-off-by: Aubin Constans <aubin.constans@microchip.com>