target/sam9x60: add system addresses of ROMs for PMECC tables
When the GALOIS_TABLE_{512|1024}_ROM_MAPPING macros are not defined, the
memory footprint of the PMECC driver is increased by 192KiB because the
PMEEC tables are then included to the driver as read-only data to be
relocated into the system RAM, whereas those PMEEC data are supposed to be
read from the dedicated ROMs when available.

This patch allows the 'nandflash' applet to fit once again into SRAM0.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>