scripts: on Windows hosts debug through a Unix adapter
Prepend the gdb command with a Unix adapter, provided such adapter is
installed and found in the PATH. From then, CTRL-C signals, instead of
being caught by MinTTY pty or Cygwin sshd, are handled by gdb.

This allows the user to break into the program.

The test relies on 'command' which is a portable, Bash builtin command.
We do not introduce any extra mandatory dependency here, given that the
Software Package depends on Bash already.

The Unix adapter is optional. Find more about it at
<https://github.com/rprichard/winpty>.

Signed-off-by: Aubin Constans <aubin.constans@microchip.com>