drivers/nvm/otp: remove 'payload_offset' variable in otp_read_packet()
The value 'payload_size - (payload_offset * 4)' is a loop invariant for
the while () loop.

Hence, the value written in '*actually_read' after the last interation of
the while () loop, 'payload_offset * 4', is actually the value of
'payload_size' before the first iteration of this loop. So we don't need
to update the 'payload_offset' variable within the loop.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>