drivers/nvm/otp: prepare to get rid of 'union otp_packet_header'
The OTPC driver API should not expose the internal layout of the OTPC_HR
register. Indeed, if the layout of this register is updated in further
versions of the OTPC hardware then the driver API would be broken.

Next step will be to change the prototype of the otp_write_packet()
function.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>