drivers/nvm/otp: remove dummy 'const' on integer parameter
The 'const' keyword makes sense and should be used whenever possible for
pointer and reference parameters of functions as it provides valuable
information to the caller of those functions: the data pointed to or
referenced by the function parameter won't be modified, hence no side
effect.

However the 'const' keyword is useless on integer parameter: it gives no
information to the caller but prevents the parameter to be updated
within the function body, which anyway as no side effect for the caller.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>