drivers/nvm/otp: fix race in otp_update_payload()
Polling for the PGM bit in OTPC_SR is risky and may lead to a race if
the driver starts to poll on this bit when the controller has already
completed the program operation (hence the PGM bit is 0 again in OTPC_SR).

On the other hand, the EOP (End Of Program) bit in OTP_ISR is cleared on
read, so it is safe to poll on this register.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>