samba_applets/common: remove auto-detection of the external xtal frequency
When called with a 0 parameter, pmc_set_main_oscillator_freq() triggers
a detection of the external crystal to guess its frequency. Doing so,
pmc_set_main_oscillator_freq() may change the clock tree, which is not
very well suited for sam-ba applets.

Actually, when the USB PLL has not been enabled by the ROM code, the
internal 12MHz RC is still selected as the source of the MAIN clock.
Hence, we don't need to bother about the external crystal at all.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>