target/sama5d2,d4: grant non-secure access to the entire DDR AESB CS address space
DDR2 Port 0 is the port connected the AES Bridge.
Writing data to the DDR AESB Chip Select address space is the way to
write data to external RAM, with the AESB encrypting it on the fly and
transparently.

As done already for the other DDR2 Ports (1 to 7), in the 64-bit AHB
Matrix, configure the DDR2 Port 0 slave so that its 4 equally-sized
regions cover the full memory space of on-board DDR-SDRAM.
As an example, boards equipped with 512 MiB DDR-SDRAM configure 128 MiB
per region.

All 4 regions are entirely configured for Non-secure access.

In the end, according to the SAMA5D2 Datasheet, the 8 DDR2 Ports slaves
must be configured the same.

Before this change, DDR2 Port 0 was typically assigned 1 MiB-sized slave
regions. Deep accesses into the DDR AESB Chip Select address space
(beyond its base address plus a 4 MiB offset) were hitting the Access
Denied Area of this slave.

Signed-off-by: Aubin Constans <aubin.constans@microchip.com>