driver/nvm/nand: fix ecc correction for sama5d3 with scrambling
For sama5d3, ecc data is inverse data, we have to check all 'FF' for
erased page. when nfc sram is enabled, we can not read spare in NFC
sram, otherwise, the scrambled data in sram will be overwitten, instead,
we could only check spare 'FF' with out NFC sram.

Signed-off-by: LI BIN <bin.li@microchip.com>